requirements.txt
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.2
greenlet==1.1.2
idna==3.3
ipykernel==6.9.1
ipython==8.1.0
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
logging==0.4.9.6
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.22.2
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.28
psycopg2==2.9.3
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.11.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
stack-data==0.2.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
urllib3==1.26.8
wcwidth==0.2.5
